feat: simplify list of widgets
