--- apps/platform/src/pages/DrugPage/ProfileWidgets.ts
@@ -1,32 +1,0 @@
-import { summaryUtils } from 'ui';
-import { Drug, Widget } from 'sections';
-import { gql } from '@apollo/client';
-import ProfileHeader from "./ProfileHeader";
-
-// This is the order of the widgets in the profile page
-export const drugProfileWidgets = new Map<string, Widget>([
-  [Drug.MechanismsOfAction.definition.id, Drug.MechanismsOfAction],
-  [Drug.Indications.definition.id, Drug.Indications],
-  [Drug.KnownDrugs.definition.id, Drug.KnownDrugs],
-  [Drug.DrugWarnings.definition.id, Drug.DrugWarnings],
-  [Drug.Pharmacogenomics.definition.id, Drug.Pharmacogenomics],
-  [Drug.AdverseEvents.definition.id, Drug.AdverseEvents],
-  [Drug.Bibliography.definition.id, Drug.Bibliography],
-]);
-
-export const drugProfileWidgetsSummaries = Array.from(drugProfileWidgets.values()).map(
-  (widget) => widget.Summary
-);
-
-const DRUG_PROFILE_SUMMARY_FRAGMENT = summaryUtils.createSummaryFragment(drugProfileWidgetsSummaries, "Drug");
-export const DRUG_PROFILE_QUERY = gql`
-  query DrugProfileQuery($chemblId: String!) {
-    drug(chemblId: $chemblId) {
-      id
-      ...DrugProfileHeaderFragment
-      ...DrugProfileSummaryFragment
-    }
-  }
-  ${ProfileHeader.fragments.profileHeader}
-  ${DRUG_PROFILE_SUMMARY_FRAGMENT}
-`;
